+++ src/app/api/admin/students/[id]/route.ts
@@ -1,6 +1,7 @@
-import { NextResponse } from "next/server";
+import { requireAdmin } from "@/guards/requireAdmin";
 import { getStudentByStudentId } from "@/services/student.service";
 import { deleteStudentById } from "@/controllers/student.controller";
+import { NextResponse } from "next/server";
 
 type Params = { id: string };
 
@@ -8,7 +9,9 @@ export async function GET(
   _: Request,
   context: { params: Promise<Params> }
 ) {
+  await requireAdmin();
   const { id } = await context.params;
+
   const student = await getStudentByStudentId(id);
   return NextResponse.json({ student });
 }
@@ -17,7 +20,9 @@ export async function DELETE(
   _: Request,
   context: { params: Promise<Params> }
 ) {
-  const { id } = await context.params; 
+  await requireAdmin();
+  const { id } = await context.params;
+
   await deleteStudentById(id);
   return NextResponse.json({ message: "Student deleted" });
 }
